Copy keys from local to sync storage on upgrade.
This commit introduces a versioning mechanism, storing a
version string to a '_version' key in sync storage. On
upgrade from 1.0.2, all keys are copied from localStorage to
chrome.storage.sync and a version key is added to the
storage. Assuming the version key is successfully written,
this only occcurs once and is attempted each time common.js
is loaded.